Safari-609.1.11.1/LayoutTests/fast/canvas/webgl/webgl-compressed-texture-astc-expected.txt

This test verifies the functionality of the WEBGL_compressed_texture_astc extension, if it is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS WebGL context exists
Testing binding enum with extension disabled
PASS supportedFormats is []

PASS No WEBGL_compressed_texture_astc support -- this is legal
PASS WEBGL_compressed_texture_astc not listed as supported and getExtension failed -- this is legal

PASS successfullyParsed is true

TEST COMPLETE
